Now a fancy allocator
