+++ wasm-calc8/src/Calc/Repl.hs
@@ -18,10 +18,10 @@ import Calc.Linearity
 import Calc.Parser
 import Calc.Parser.Types
 import Calc.Typecheck
-import Calc.Wasm.FromExpr
+import Calc.Wasm.FromExpr.Expr
 import Calc.Wasm.Run
-import Calc.Wasm.ToWasm
-import Calc.Wasm.Types
+import Calc.Wasm.ToWasm.Expr
+import Calc.Wasm.ToWasm.Types
 import Control.Monad.IO.Class
 import Data.Text (Text)
 import qualified Data.Text as T
